[BUGFIX] Fix phpstan errors from ReturnTypeRule
EXT:impexp
EXT:indexed_search
EXT:info
EXT:install

Resolves: #94264
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69394
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>

+++ typo3/sysext/indexed_search/Classes/Utility/IndexedSearchUtility.php
@@ -47,7 +47,7 @@ class IndexedSearchUtility
      */
     public static function md5inthash($stringToHash)
     {
-        return hexdec(substr(md5($stringToHash), 0, 7));
+        return (int)hexdec(substr(md5($stringToHash), 0, 7));
     }
 
     /**
@@ -174,6 +174,6 @@ class IndexedSearchUtility
      */
     public static function milliseconds()
     {
-        return round(microtime(true) * 1000);
+        return (int)round(microtime(true) * 1000);
     }
 }
